Pooling connections for writes.
Started using a pool of connections to deal with multiple sequentials
writes to a single socket, this avoids a `short error` failures that was
happening too easy.

Some minor changes are a encoder/decoder instead of dealing with the bytes directly
from the bufio. Treating the shutdown better, now will wait for the
blocking `accept` call to also finish before returning. Moved
*everything* to the public package because dealing with dependecy cycles
is a pain. At last added a simple benchmark, but not sure yet if it is
right?

Some of possible improvements should focus on:

* Worker pool for writes;
* Handling idle connections;