Add message buffer.
Previously, if nobody is listening the produced datagrams, the write
operation will block the `network_connection` of receiving more
messages. To avoid this, a `buffer` was created, that will hold all
received messages into a FIFO queue.

The first action is trying to deliver the message through the channel,
if is not possible it will be added to the message buffer to be
delivered later. This behavior can lead to message being delivered out
of order, but I think is something simple to solve.

Another problem is that messages can be stuck on the buffer forever, and
the buffer will only grow. If we do not care for discarding messages, a
TTL can be added with a configurable parameter.

Also fixed the benchmark, previously we were waiting to close after 1
second without receiving, and this was breaking the `ns/ops` the always
be 1 second. Now we only validate the sending of 1024 messages, since a
single message made the `b.N` to execute too many times, sometimes not
finishing.

Executing the benchmarks without consuming the messages also lead to
benchmarks executing forever. I think is because the use of mutexes to
insert the messages to the buffer.